+++ EpicJewels/EffectHelpers/SkillIncrease.cs
@@ -25,31 +25,35 @@ namespace EpicJewels.EffectHelpers
             switch (selectedSkill)
             {
                 case SkillType.Swords:
-                    if (player.GetEffectPower<GemEffects.ExpertSwordsman.Config>("ExpertSwordsman").Power > 0)
+                    if (player.GetEffectPower<GemEffects.ExpertSwordsman.Config>("ExpertSwordsman").Power > 0 || player.GetEffectPower<GemEffects.WeaponMaster.Config>("WeaponMaster").Power > 0)
                     {
                         skill_increase += player.GetEffectPower<GemEffects.ExpertSwordsman.Config>("ExpertSwordsman").Power;
-                        EpicJewels.EJLog.LogDebug($"ExpertSwordsman skill increase {skill_increase}");
+                        skill_increase += player.GetEffectPower<GemEffects.WeaponMaster.Config>("WeaponMaster").Power;
+                        EpicJewels.EJLog.LogDebug($"Swords skill increase {skill_increase}");
                     }
                     break;
                 case SkillType.Spears:
-                    if (player.GetEffectPower<GemEffects.ExpertSpearmaiden.Config>("ExpertSpearmaiden").Power > 0)
+                    if (player.GetEffectPower<GemEffects.ExpertSpearmaiden.Config>("ExpertSpearmaiden").Power > 0 || player.GetEffectPower<GemEffects.WeaponMaster.Config>("WeaponMaster").Power > 0)
                     {
                         skill_increase += player.GetEffectPower<GemEffects.ExpertSpearmaiden.Config>("ExpertSpearmaiden").Power;
-                        EpicJewels.EJLog.LogDebug($"ExpertSpearmaiden skill increase {skill_increase}");
+                        skill_increase += player.GetEffectPower<GemEffects.WeaponMaster.Config>("WeaponMaster").Power;
+                        EpicJewels.EJLog.LogDebug($"Spears skill increase {skill_increase}");
                     }
                     break;
                 case SkillType.Clubs:
-                    if (player.GetEffectPower<GemEffects.ExpertSmasher.Config>("ExpertSmasher").Power > 0)
+                    if (player.GetEffectPower<GemEffects.ExpertSmasher.Config>("ExpertSmasher").Power > 0 || player.GetEffectPower<GemEffects.WeaponMaster.Config>("WeaponMaster").Power > 0)
                     {
                         skill_increase += player.GetEffectPower<GemEffects.ExpertSmasher.Config>("ExpertSmasher").Power;
-                        EpicJewels.EJLog.LogDebug($"ExpertSmasher skill increase {skill_increase}");
+                        skill_increase += player.GetEffectPower<GemEffects.WeaponMaster.Config>("WeaponMaster").Power;
+                        EpicJewels.EJLog.LogDebug($"Clubs skill increase {skill_increase}");
                     }
                     break;
                 case SkillType.Polearms:
-                    if (player.GetEffectPower<GemEffects.ExpertPolearms.Config>("ExpertPolearms").Power > 0)
+                    if (player.GetEffectPower<GemEffects.ExpertPolearms.Config>("ExpertPolearms").Power > 0 || player.GetEffectPower<GemEffects.WeaponMaster.Config>("WeaponMaster").Power > 0)
                     {
                         skill_increase += player.GetEffectPower<GemEffects.ExpertPolearms.Config>("ExpertPolearms").Power;
-                        EpicJewels.EJLog.LogDebug($"ExpertPolearms skill increase {skill_increase}");
+                        skill_increase += player.GetEffectPower<GemEffects.WeaponMaster.Config>("WeaponMaster").Power;
+                        EpicJewels.EJLog.LogDebug($"Polearms skill increase {skill_increase}");
                     }
                     break;
                 case SkillType.BloodMagic:
@@ -57,49 +61,45 @@ namespace EpicJewels.EffectHelpers
                     if (player.GetEffectPower<GemEffects.ExpertMage.Config>("ExpertMage").Power > 0)
                     {
                         skill_increase += player.GetEffectPower<GemEffects.ExpertMage.Config>("ExpertMage").Power;
-                        EpicJewels.EJLog.LogDebug($"ExpertMage skill increase {skill_increase}");
+                        EpicJewels.EJLog.LogDebug($"Magic skill increase {skill_increase}");
                     }
                     break;
                 case SkillType.Fishing:
                     if (player.GetEffectPower<GemEffects.ExpertFisher.Config>("ExpertFisher").Power > 0)
                     {
                         skill_increase += player.GetEffectPower<GemEffects.ExpertFisher.Config>("ExpertFisher").Power;
-                        EpicJewels.EJLog.LogDebug($"ExpertFisher skill increase {skill_increase}");
+                        EpicJewels.EJLog.LogDebug($"Fishing skill increase {skill_increase}");
                     }
                     break;
                 case SkillType.Knives:
-                    if (player.GetEffectPower<GemEffects.ExpertDaggers.Config>("ExpertDaggers").Power > 0)
+                    if (player.GetEffectPower<GemEffects.ExpertDaggers.Config>("ExpertDaggers").Power > 0 || player.GetEffectPower<GemEffects.WeaponMaster.Config>("WeaponMaster").Power > 0)
                     {
                         skill_increase += player.GetEffectPower<GemEffects.ExpertDaggers.Config>("ExpertDaggers").Power;
-                        EpicJewels.EJLog.LogDebug($"ExpertDaggers skill increase {skill_increase}");
+                        skill_increase += player.GetEffectPower<GemEffects.WeaponMaster.Config>("WeaponMaster").Power;
+                        EpicJewels.EJLog.LogDebug($"Knives skill increase {skill_increase}");
                     }
                     break;
                 case SkillType.Unarmed:
-                    if (player.GetEffectPower<GemEffects.ExpertBrawler.Config>("ExpertBrawler").Power > 0)
+                    if (player.GetEffectPower<GemEffects.ExpertBrawler.Config>("ExpertBrawler").Power > 0 || player.GetEffectPower<GemEffects.WeaponMaster.Config>("WeaponMaster").Power > 0)
                     {
                         skill_increase += player.GetEffectPower<GemEffects.ExpertBrawler.Config>("ExpertBrawler").Power;
-                        EpicJewels.EJLog.LogDebug($"ExpertBrawler skill increase {skill_increase}");
+                        skill_increase += player.GetEffectPower<GemEffects.WeaponMaster.Config>("WeaponMaster").Power;
+                        EpicJewels.EJLog.LogDebug($"Unarmed skill increase {skill_increase}");
                     }
                     break;
                 case SkillType.Axes:
-                    if (player.GetEffectPower<GemEffects.ExpertAxemaster.Config>("ExpertAxemaster").Power > 0)
+                    if (player.GetEffectPower<GemEffects.ExpertAxemaster.Config>("ExpertAxemaster").Power > 0 || player.GetEffectPower<GemEffects.WeaponMaster.Config>("WeaponMaster").Power > 0)
                     {
                         skill_increase += player.GetEffectPower<GemEffects.ExpertAxemaster.Config>("ExpertAxemaster").Power;
-                        EpicJewels.EJLog.LogDebug($"ExpertAxemaster skill increase {skill_increase}");
-                    }
-                    break;
-                case SkillType.Blocking:
-                    if (player.GetEffectPower<GemEffects.ExpertShieldbearer.Config>("ExpertShieldbearer").Power > 0)
-                    {
-                        skill_increase += player.GetEffectPower<GemEffects.ExpertShieldbearer.Config>("ExpertShieldbearer").Power;
-                        EpicJewels.EJLog.LogDebug($"ExpertShieldbearer skill increase {skill_increase}");
+                        skill_increase += player.GetEffectPower<GemEffects.WeaponMaster.Config>("WeaponMaster").Power;
+                        EpicJewels.EJLog.LogDebug($"Axes skill increase {skill_increase}");
                     }
                     break;
                 case SkillType.Jump:
                     if (player.GetEffectPower<GemEffects.ExpertAcrobat.Config>("ExpertAcrobat").Power > 0)
                     {
                         skill_increase += player.GetEffectPower<GemEffects.ExpertAcrobat.Config>("ExpertAcrobat").Power;
-                        EpicJewels.EJLog.LogDebug($"ExpertAcrobat skill increase {skill_increase}");
+                        EpicJewels.EJLog.LogDebug($"Jump skill increase {skill_increase}");
                     }
                     break;
                 case SkillType.WoodCutting:
@@ -107,46 +107,57 @@ namespace EpicJewels.EffectHelpers
                     if (player.GetEffectPower<GemEffects.ExpertHarvester.Config>("ExpertHarvester").Power > 0)
                     {
                         skill_increase += player.GetEffectPower<GemEffects.ExpertHarvester.Config>("ExpertHarvester").Power;
-                        EpicJewels.EJLog.LogDebug($"ExpertHarvester skill increase {skill_increase}");
+                        EpicJewels.EJLog.LogDebug($"Pickaxe & woodcutting skill increase {skill_increase}");
                     }
                     break;
                 case SkillType.Run:
                     if (player.GetEffectPower<GemEffects.ExpertSprinter.Config>("ExpertSprinter").Power > 0)
                     {
                         skill_increase += player.GetEffectPower<GemEffects.ExpertSprinter.Config>("ExpertSprinter").Power;
-                        EpicJewels.EJLog.LogDebug($"ExpertSprinter skill increase {skill_increase}");
+                        EpicJewels.EJLog.LogDebug($"Run skill increase {skill_increase}");
+                    }
+                    break;
+                case SkillType.Crossbows:
+                    if (player.GetEffectPower<GemEffects.WeaponMaster.Config>("WeaponMaster").Power > 0)
+                    {
+                        skill_increase += player.GetEffectPower<GemEffects.WeaponMaster.Config>("WeaponMaster").Power;
+                        EpicJewels.EJLog.LogDebug($"Crossbow skill increase {skill_increase}");
                     }
                     break;
                 // These empty ones are already provided by base jewelcrafting
+                case SkillType.Blocking:
                 case SkillType.None:
-                case SkillType.Crossbows:
+
                 case SkillType.Bows:
                 case SkillType.Sneak:
                 case SkillType.Swim:
                 case SkillType.Ride:
                     break;
                 case SkillType.All:
+                    float wm_bonus = player.GetEffectPower<GemEffects.WeaponMaster.Config>("WeaponMaster").Power;
                     // harvester is included twice once for pickaxe, once for woodcutting
                     skill_increase += player.GetEffectPower<GemEffects.ExpertHarvester.Config>("ExpertHarvester").Power;
                     skill_increase += player.GetEffectPower<GemEffects.ExpertHarvester.Config>("ExpertHarvester").Power;
-                    skill_increase += player.GetEffectPower<GemEffects.ExpertSwordsman.Config>("ExpertSwordsman").Power;
-                    skill_increase += player.GetEffectPower<GemEffects.ExpertSpearmaiden.Config>("ExpertSpearmaiden").Power;
-                    skill_increase += player.GetEffectPower<GemEffects.ExpertSmasher.Config>("ExpertSmasher").Power;
-                    skill_increase += player.GetEffectPower<GemEffects.ExpertPolearms.Config>("ExpertPolearms").Power;
+                    skill_increase += wm_bonus + player.GetEffectPower<GemEffects.ExpertSwordsman.Config>("ExpertSwordsman").Power;
+                    skill_increase += wm_bonus + player.GetEffectPower<GemEffects.ExpertSpearmaiden.Config>("ExpertSpearmaiden").Power;
+                    skill_increase += wm_bonus + player.GetEffectPower<GemEffects.ExpertSmasher.Config>("ExpertSmasher").Power;
+                    skill_increase += wm_bonus + player.GetEffectPower<GemEffects.ExpertPolearms.Config>("ExpertPolearms").Power;
                     // mage is included twice, one for blood once for elemental
                     skill_increase += player.GetEffectPower<GemEffects.ExpertMage.Config>("ExpertMage").Power;
                     skill_increase += player.GetEffectPower<GemEffects.ExpertMage.Config>("ExpertMage").Power;
                     skill_increase += player.GetEffectPower<GemEffects.ExpertFisher.Config>("ExpertFisher").Power;
-                    skill_increase += player.GetEffectPower<GemEffects.ExpertDaggers.Config>("ExpertDaggers").Power;
-                    skill_increase += player.GetEffectPower<GemEffects.ExpertBrawler.Config>("ExpertBrawler").Power;
-                    skill_increase += player.GetEffectPower<GemEffects.ExpertAxemaster.Config>("ExpertAxemaster").Power;
-                    skill_increase += player.GetEffectPower<GemEffects.ExpertShieldbearer.Config>("ExpertShieldbearer").Power;
+                    skill_increase += wm_bonus + player.GetEffectPower<GemEffects.ExpertDaggers.Config>("ExpertDaggers").Power;
+                    skill_increase += wm_bonus + player.GetEffectPower<GemEffects.ExpertBrawler.Config>("ExpertBrawler").Power;
+                    skill_increase += wm_bonus + player.GetEffectPower<GemEffects.ExpertAxemaster.Config>("ExpertAxemaster").Power;
+                    // skill_increase += wm_bonus // Shield bonus
                     skill_increase += player.GetEffectPower<GemEffects.ExpertAcrobat.Config>("ExpertAcrobat").Power;
+
+                    EpicJewels.EJLog.LogDebug($"All skills increase {skill_increase}");
                     break;
             }
 
             float final_skill_increase = (skill_increase + 100) / 100;
-            EpicJewels.EJLog.LogDebug($"Total skill increase {final_skill_increase}");
+            // EpicJewels.EJLog.LogDebug($"Total skill increase {final_skill_increase}");
             return (current_skill_level * final_skill_increase);
         }