+++ EpicJewels/EpicJewels.cs
@@ -1,7 +1,15 @@
 ﻿using BepInEx;
 using BepInEx.Logging;
 using HarmonyLib;
+using System.IO;
+using System.Linq;
+using System;
 using System.Reflection;
+using UnityEngine;
+using YamlDotNet.Serialization.NamingConventions;
+using YamlDotNet.Serialization;
+using LocalizationManager;
+using Mono.Cecil;
 
 namespace EpicJewels
 {
@@ -11,11 +19,16 @@ namespace EpicJewels
     {
         public const string PluginGUID = "MidnightsFX.EpicJewels";
         public const string PluginName = "EpicJewels";
-        public const string PluginVersion = "0.2.5";
+        public const string PluginVersion = "0.2.6";
 
         public static readonly ManualLogSource EJLog = BepInEx.Logging.Logger.CreateLogSource(PluginName);
 
-        public Common.Config cfg;
+        internal Common.Config cfg;
+        internal static AssetBundle EmbeddedResourceBundle;
+        internal static Harmony Harmony = new Harmony(PluginGUID);
+        public static IDeserializer yamldeserializer = new DeserializerBuilder().WithNamingConvention(CamelCaseNamingConvention.Instance).Build();
+        public static ISerializer yamlserializer = new SerializerBuilder().WithNamingConvention(CamelCaseNamingConvention.Instance).DisableAliases().Build();
+        public static Material spiritCreature;
 
         // Use this class to add your own localization to the game
         // https://valheim-modding.github.io/Jotunn/tutorials/localization.html
@@ -23,16 +36,60 @@ namespace EpicJewels
 
         private void Awake()
         {
+            Localizer.Load();
             cfg = new Common.Config(Config);
-            // Jotunn comes with its own Logger class to provide a consistent Log style for all mods using it
+            EmbeddedResourceBundle = LoadAssetBundle("EpicJewels.AssetsEmbedded.epicjewels");
+            EJLog.LogDebug("Logging embedded assets.");
+            foreach (string asset_name in EmbeddedResourceBundle.GetAllAssetNames())
+            {
+                EJLog.LogDebug(asset_name);
+            }
             EJLog.LogInfo("Let the gems flow.");
+            // Common.Translations.Instance.LoadLocalizations();
             GemEffects.EffectList.AddGemEffects();
-            CustomGems.AddGems();
+            new CustomGems(EmbeddedResourceBundle);
+
+            spiritCreature = EmbeddedResourceBundle.LoadAsset<Material>("assets/custom/fbx_exports/gems/spirit_animal_mat.mat");
 
             Assembly assembly = Assembly.GetExecutingAssembly();
-            Harmony harmony = new(PluginGUID);
-            harmony.PatchAll(assembly);
+            Harmony.PatchAll(assembly);
+        }
 
+        public static AssetBundle LoadAssetBundle(string bundleName)
+        {
+            var resourceAssembly = typeof(EpicJewels).Assembly;
+            string text = null;
+            AssetBundle result;
+            try
+            {
+                text = resourceAssembly.GetManifestResourceNames().Single((string str) => str.EndsWith(bundleName));
+            }
+            catch (Exception) {}
+            using (Stream stream = resourceAssembly.GetManifestResourceStream(text))
+            {
+                result = AssetBundle.LoadFromStream(stream);
+            }
+            return result;
+        }
+
+        public static String LoadEmbeddedAssetToString(string assetName) 
+        {
+            var resourceAssembly = typeof(EpicJewels).Assembly;
+            string text = null;
+            string result;
+            try
+            {
+                text = resourceAssembly.GetManifestResourceNames().Single((string str) => str.EndsWith(assetName));
+            } catch (Exception) { }
+            if (text == null) { return null; }
+            using (Stream stream = resourceAssembly.GetManifestResourceStream(text))
+            {
+                using (StreamReader reader = new StreamReader(stream))
+                {
+                    result = reader.ReadToEnd();
+                }
+            }
+            return result;
         }
     }
 }