+++ EpicJewels/GemEffects/IntenseFire.cs
@@ -1,0 +1,16 @@
+﻿using JetBrains.Annotations;
+using Jewelcrafting;
+
+namespace EpicJewels.GemEffects
+{
+    public static class IntenseFire
+    {
+        [PublicAPI]
+        public struct Config
+        {
+            [AdditivePowerAttribute] public float Power;
+        }
+    }
+
+
+}