+++ EpicJewels/GemEffects/WaterSwiftness.cs
@@ -18,9 +18,11 @@ namespace EpicJewels.GemEffects
         {
             private static void Postfix(Player __instance, ref float __result)
             {
-                if (__instance.GetSEMan().HaveStatusEffect(wetstatus) && __instance.GetEffectPower<Config>("WaterSwiftness").Power > 0)
+                if (__instance.GetSEMan().HaveStatusEffect(wetstatus) && (__instance.GetEffectPower<Config>("WaterSwiftness").Power > 0 || __instance.GetEffectPower<Config>("SlipperyWhenWet").Power > 0))
                 {
-                    __result *= ((__instance.GetEffectPower<Config>("WaterSwiftness").Power + 100f) / 100f);
+                    float speed_modifier = ((__instance.GetEffectPower<Config>("WaterSwiftness").Power + __instance.GetEffectPower<Config>("SlipperyWhenWet").Power + 100f) / 100f);
+                    // EpicJewels.EJLog.LogDebug($"Wet jog speed modifier: {speed_modifier}");
+                    __result *= speed_modifier;
                 }
                 
             }
@@ -31,9 +33,11 @@ namespace EpicJewels.GemEffects
         {
             private static void Postfix(Player __instance, ref float __result)
             {
-                if (__instance.GetSEMan().HaveStatusEffect(wetstatus) && __instance.GetEffectPower<Config>("WaterSwiftness").Power > 0)
+                if (__instance.GetSEMan().HaveStatusEffect(wetstatus) && (__instance.GetEffectPower<Config>("WaterSwiftness").Power > 0 || __instance.GetEffectPower<Config>("SlipperyWhenWet").Power > 0))
                 {
-                    __result *= ((__instance.GetEffectPower<Config>("WaterSwiftness").Power + 100f) / 100f);
+                    float speed_modifier = ((__instance.GetEffectPower<Config>("WaterSwiftness").Power + __instance.GetEffectPower<Config>("SlipperyWhenWet").Power + 100f) / 100f);
+                    // EpicJewels.EJLog.LogDebug($"Wet run speed modifier: {speed_modifier}");
+                    __result *= speed_modifier;
                 }
             }
         }